// src/app/api/models/comment-aggregate.ts
/* tslint:disable */
export interface CommentAggregate {
  liked?: boolean;
  commentId?: number;
  completedChallengeId?: number;
  createdDate?: string;
  description?: string;
  firstName?: string;
  lastName?: string;
  commitedActivityId?: number;
  noOfLoves?: number;
  noOfReplies?: number;
  profilePicture?: string;
  profilePictureContentType?: string;
  timeElapsed?: string;
  userId?: string;
}